requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
langchain==0.1.0
langchain-google-genai==0.0.8
langchain-community==0.0.13
chromadb==0.4.22
sentence-transformers==2.2.2
pypdf2==3.0.1
python-dotenv==1.0.0
google-generativeai==0.3.2
gunicorn==21.2.0
Werkzeug==2.3.7